Biological Year Definition Tab
After selecting a dropdown for the start month (which begins from october), you'll be prompted to enter seasons for each month in the year.

The displayed biological year's start month will change based on the dropdown.

Information entered for the biological year will be saved in a .csv file which can be found in Data/Settings