CleanUp, Data Select
Fixing some bugs, and cleaning up files (deleted unnecessary files)

Data Select: Not tested yet, but this is a new tab in the Data Define tab, where the user will be retrieving the data for visuals.
In this tab, the user will choose what rasterLayer data (.tif files) to extract with the migration data, and clicking the "retrieve" button will retrieve the data for later visualizations.

There might be a better way for this, but since I didn't want to call the retrieve command every time a new visual was requested (that would be much slower and unnecessary) I created a tab for it.

Can't test the Data Select tab yet because I don't have .tif files to test the data extraction with