Select Data Tab
Tab is located in Data Define --> Retrieve Data.
Select the different TIF File information, and click "Load Data". The data will be loaded into the migrationData variable, to then later be used in visualizations.
Because I see that there are so many different data types for the TIF files, I want to make it so you can select what data you want to load, as to try to optimize time (especially if you only want to view one tif layer).

Keep in mind, in order to update the data in memory (say, you change the biological year or something), you'll need to go to this tab to reload the data.

I haven't tested the date ranges with the layers yet. I'm also not sure how correct the raster::extract() function is, with the layers; will need more testing.